# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API de Pedidos - Desafio Jitterbit",
    "version": "1.0.0",
    "description": "API para gerenciamento de pedidos"
  },
  "servers": [
    { "url": "http://localhost:3000" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "OrderInput": {
        "type": "object",
        "required": ["numeroPedido", "valorTotal", "dataCriacao", "items"],
        "properties": {
          "numeroPedido": { "type": "string", "example": "v10089015vdb-01" },
          "valorTotal": { "type": "number", "example": 10000 },
          "dataCriacao": { "type": "string", "example": "2023-07-19T12:24:11.5299601+00:00" },
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "idItem": { "type": "string", "example": "2434" },
                "quantidadeItem": { "type": "integer", "example": 1 },
                "valorItem": { "type": "number", "example": 1000 }
              }
            }
          }
        }
      },
      "Order": {
        "type": "object",
        "properties": {
          "orderId": { "type": "string" },
          "value": { "type": "number" },
          "creationDate": { "type": "string" },
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "productId": { "type": "integer" },
                "quantity": { "type": "integer" },
                "price": { "type": "number" }
              }
            }
          }
        }
      }
    }
  },
  "paths": {
    "/auth/login": {
      "post": {
        "summary": "Autenticação - Gerar token JWT",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": { "type": "string", "example": "admin" },
                  "password": { "type": "string", "example": "admin" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Token gerado com sucesso" },
          "401": { "description": "Credenciais inválidas" }
        }
      }
    },
    "/order": {
      "post": {
        "summary": "Criar um novo pedido",
        "tags": ["Pedidos"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/OrderInput" }
            }
          }
        },
        "responses": {
          "201": { "description": "Pedido criado com sucesso" },
          "400": { "description": "Campos obrigatórios faltando" },
          "401": { "description": "Token não fornecido ou inválido" },
          "409": { "description": "Pedido já existe" }
        }
      }
    },
    "/order/list": {
      "get": {
        "summary": "Listar todos os pedidos",
        "tags": ["Pedidos"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Lista de pedidos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Order" }
                }
              }
            }
          }
        }
      }
    },
    "/order/{orderId}": {
      "get": {
        "summary": "Obter pedido pelo número",
        "tags": ["Pedidos"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Dados do pedido" },
          "404": { "description": "Pedido não encontrado" }
        }
      },
      "put": {
        "summary": "Atualizar pedido",
        "tags": ["Pedidos"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/OrderInput" }
            }
          }
        },
        "responses": {
          "200": { "description": "Pedido atualizado" },
          "404": { "description": "Pedido não encontrado" }
        }
      },
      "delete": {
        "summary": "Deletar pedido",
        "tags": ["Pedidos"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Pedido deletado" },
          "404": { "description": "Pedido não encontrado" }
        }
      }
    }
  }
}